[flue] Extract code-review comment rendering into lib/code-review-render.ts
